Forumla One Team Information Er Databse Model
The PDF contains paragraphs written in simple English explaining the ER model to the user. Additionally, it includes three types of queries at the end, totaling six different queries. These are provided as examples to help the user better understand and utilize the information.Additionally, the attached visual on paper is a concept illustrating the ER model of the system.